Change to 44.1kHz and accept integer input
WASAPI does not resample for us, and some of the machines
that are supposed to run this won't provide 48kHz float.

In the future, we should try to allow both 44.1 and 48kHz
to make optimal use of more hardware.

For now, we still require floating point output, but
contemporary sound architectures are more willing to resample
and repack audio output, which means we are probably fine.

By the way, we can now correctly select raw and non-raw devices.